move DeviceSecurityVariation into types
