+++ src/components/views/settings/devices/filter.ts
@@ -14,13 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import { DeviceWithVerification } from "./types";
-
-export enum DeviceSecurityVariation {
-    Verified = 'Verified',
-    Unverified = 'Unverified',
-    Inactive = 'Inactive',
-}
+import { DeviceWithVerification, DeviceSecurityVariation } from "./types";
 
 type DeviceFilterCondition = (device: DeviceWithVerification) => boolean;
