+++ src/components/views/settings/devices/types.ts
@@ -18,3 +18,9 @@ import { IMyDevice } from "matrix-js-sdk/src/matrix";
 
 export type DeviceWithVerification = IMyDevice & { isVerified: boolean | null };
 export type DevicesDictionary = Record<DeviceWithVerification['device_id'], DeviceWithVerification>;
+
+export enum DeviceSecurityVariation {
+    Verified = 'Verified',
+    Unverified = 'Unverified',
+    Inactive = 'Inactive',
+}
